python/django-example/requirements.txt
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.2.7
djangorestframework==3.16.1
djlint==1.36.4
EditorConfig==0.17.1
jsbeautifier==1.15.4
json5==0.12.1
pathspec==0.12.1
PyYAML==6.0.3
regex==2025.9.18
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
